+++ packages/taro-cli/src/mini/index.ts
@@ -13,7 +13,8 @@ import {
   copyFiles,
   unzip,
   shouldUseYarn,
-  shouldUseCnpm
+  shouldUseCnpm,
+  resolvePureScriptPath
 } from '../util'
 import { processTypeEnum, BUILD_TYPES } from '../util/constants'
 import { IMiniAppBuildConfig } from '../util/types'
@@ -28,12 +29,19 @@ import {
   setQuickappManifest
 } from './helper'
 import { buildEntry } from './entry'
-import { buildPages } from './page'
+import { buildPages, buildSinglePage } from './page'
 import { watchFiles } from './watch'
 import { downloadGithubRepoLatestRelease } from '../util/dowload'
+import { buildSingleComponent } from './component'
 
 function buildProjectConfig () {
   const { buildAdapter, sourceDir, outputDir, outputDirName, appPath } = getBuildData()
+
+  if (buildAdapter === BUILD_TYPES.JD) {
+    // 京东小程序暂不支持 project.config.json
+    return
+  }
+
   let projectConfigFileName = `project.${buildAdapter}.json`
   if (buildAdapter === BUILD_TYPES.WEAPP || buildAdapter === BUILD_TYPES.QQ) {
     projectConfigFileName = 'project.config.json'
@@ -228,7 +236,18 @@ async function runQuickApp (isWatch: boolean | void, buildData: IBuildData, port
   }
 }
 
-export async function build (appPath: string, { watch, adapter = BUILD_TYPES.WEAPP, envHasBeenSet = false, port, release }: IMiniAppBuildConfig) {
+export async function build (
+  appPath: string,
+  {
+    watch,
+    adapter = BUILD_TYPES.WEAPP,
+    envHasBeenSet = false,
+    port,
+    release,
+    page,
+    component
+  }: IMiniAppBuildConfig
+) {
   const buildData = envHasBeenSet ? getBuildData() : setBuildData(appPath, adapter)
   const isQuickApp = adapter === BUILD_TYPES.QUICKAPP
   let quickappJSON
@@ -246,6 +265,18 @@ export async function build (appPath: string, { watch, adapter = BUILD_TYPES.WEA
   }
   if (!isQuickApp) {
     copyFiles(appPath, buildData.projectConfig.copy)
+  }
+  if (page) {
+    const pagePath = path.resolve(appPath, page).replace(buildData.sourceDir, '')
+    await buildSinglePage(pagePath)
+    return
+  }
+  if (component) {
+    const componentPath = resolvePureScriptPath(path.resolve(appPath, component))
+    await buildSingleComponent({
+      path: componentPath
+    })
+    return
   }
   const appConfig = await buildEntry()
   setAppConfig(appConfig)